# [Enhancement] Optimize the inner-join performance of nest-loop-join when the right table is relatively small