@trueeyu 2023-11-20
https://github.com/Mergifyio backport branch-2.5